+++ CalibFormats/SiPixelObjects/src/PixelPortCardConfig.cc
@@ -368,10 +368,10 @@ PixelPortCardConfig::PixelPortCardConfig(std::string filename):
   string mthn = "[PixelPortCardConfig::PixelPortCardConfig()]\t\t    " ;
   //std::cout << __LINE__ << "]\t" << mthn << "filename:"<<filename<<std::endl;
 
-  unsigned int portcardpos=filename.find(std::string("portcard_"));
+  size_t portcardpos=filename.find(std::string("portcard_"));
   //std::cout << __LINE__ << "]\t" << mthn << "portcardpos:"<<portcardpos<<std::endl;
   assert(portcardpos!=std::string::npos);
-  unsigned int datpos=filename.find(std::string(".dat"));
+  size_t datpos=filename.find(std::string(".dat"));
   //std::cout << __LINE__ << "]\t" << mthn << "datpos:"<<datpos<<std::endl;
   assert(datpos!=std::string::npos);
   assert(datpos>portcardpos);
